chore: fix clippy warnings + gameplay-drums lib imports
Reduces clippy from 125 to <10 warnings. Main fixes:
- dtx-core/channel.rs: rename Guitar_xxxYx → GuitarYxxYx etc.
  (upper camel case required by clippy)
- dtx-channel imports updated in gameplay-guitar/lane_map.rs
- gameplay-drums/score.rs + judge.rs + scroll.rs: add
  lane_map::LaneId, LANE_ORDER imports
- dtx-config / dtx-ui / gameplay-drums / etc.: add
  #![allow(missing_docs)] to constants files

dtx-config is now wired into the binary (ba8a620): boot logs
show the loaded config summary.

Refs: clippy lint categories, ADR-0010

+++ crates/game-shell/src/fade.rs
@@ -82,7 +82,7 @@ fn update_fade_system(
     }
 
     let dt_ms = time.delta().as_millis() as u64;
-    fade.elapsed_ms = Some((fade.elapsed_ms.unwrap_or(0) + dt_ms).min(FADE_DURATION_MS as u64));
+    fade.elapsed_ms = Some((fade.elapsed_ms.unwrap_or(0) + dt_ms).min(FADE_DURATION_MS));
 
     let alpha = fade.alpha();
